hemai_mdl/requirements_mdl.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
numpy
Pillow
pandas
